refactor(ray): merge ray_executor_client.py into ray_executor.py
Combine the Ray Jobs client layer (RayJobConfig, RayJobClient,
get_ray_client) and the nemo-run Executor adapter (RayExecutor) into a
single module. They were two files only because they were authored
incrementally; semantically they are one Ray-executor unit.

- nemo_skills/pipeline/utils/ray_executor_client.py: deleted, content
  moved into ray_executor.py with section comments separating the
  client layer from the executor adapter.
- nemo_skills/pipeline/utils/exp.py: updated import to pull
  RayExecutor + RayJobConfig + get_ray_client from the merged module.
- tests/test_ray_executor_client.py → tests/test_ray_executor.py:
  renamed and import paths updated for the merged module.

Kept the file name ray_executor.py rather than ray.py to avoid any
risk of namespace shadowing with `import ray` inside the module.

Address PR feedback from @gwarmstrong on NVIDIA-NeMo/Skills#1435.
